refactor: PlaceEntity와 PlaceType을 domain.place 패키지로 이동
