+++ src/main/java/com/example/barrier_free/domain/favorite/FavoriteService.java
@@ -27,11 +27,11 @@ import com.example.barrier_free.domain.favorite.repository.FavoriteRepository;
 import com.example.barrier_free.domain.favorite.repository.WeeklyRankRepository;
 import com.example.barrier_free.domain.map.entity.Map;
 import com.example.barrier_free.domain.map.repository.MapRepository;
+import com.example.barrier_free.domain.place.enums.PlaceType;
 import com.example.barrier_free.domain.report.entity.Report;
 import com.example.barrier_free.domain.report.repository.ReportRepository;
 import com.example.barrier_free.domain.user.UserRepository;
 import com.example.barrier_free.domain.user.entity.User;
-import com.example.barrier_free.global.common.PlaceType;
 import com.example.barrier_free.global.exception.CustomException;
 import com.example.barrier_free.global.jwt.JwtUserUtils;
 import com.example.barrier_free.global.response.ErrorCode;
