+++ src/main/java/com/example/barrier_free/domain/favorite/dto/FavoriteRequestDto.java
@@ -1,6 +1,6 @@
 package com.example.barrier_free.domain.favorite.dto;
 
-import com.example.barrier_free.global.common.PlaceType;
+import com.example.barrier_free.domain.place.enums.PlaceType;
 
 import lombok.Getter;
