+++ src/main/java/com/example/barrier_free/domain/favorite/dto/FavoriteResponse.java
@@ -2,8 +2,8 @@ package com.example.barrier_free.domain.favorite.dto;
 
 import java.util.List;
 
+import com.example.barrier_free.domain.place.enums.PlaceType;
 import com.example.barrier_free.global.common.Place;
-import com.example.barrier_free.global.common.PlaceType;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
