+++ src/main/java/com/example/barrier_free/domain/place/controller/PlaceController.java
@@ -16,8 +16,8 @@ import com.example.barrier_free.domain.place.dto.PlaceMapMarkerResponse;
 import com.example.barrier_free.domain.place.dto.PlaceSearchCondition;
 import com.example.barrier_free.domain.place.dto.PlaceSearchResponsePage;
 import com.example.barrier_free.domain.place.dto.PlaceSummaryResponse;
+import com.example.barrier_free.domain.place.enums.PlaceType;
 import com.example.barrier_free.domain.place.service.PlaceService;
-import com.example.barrier_free.global.common.PlaceType;
 import com.example.barrier_free.global.response.ApiResponse;
 import com.example.barrier_free.global.response.SuccessCode;
