+++ src/main/java/com/example/barrier_free/domain/place/dto/PlaceMapMarkerResponse.java
@@ -1,6 +1,6 @@
 package com.example.barrier_free.domain.place.dto;
 
-import com.example.barrier_free.global.common.PlaceType;
+import com.example.barrier_free.domain.place.enums.PlaceType;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
