+++ src/main/java/com/example/barrier_free/domain/place/dto/PlaceSummaryResponse.java
@@ -2,7 +2,7 @@ package com.example.barrier_free.domain.place.dto;
 
 import java.util.List;
 
-import com.example.barrier_free.global.common.PlaceType;
+import com.example.barrier_free.domain.place.enums.PlaceType;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
