+++ src/main/java/com/example/barrier_free/global/common/Place.java
@@ -3,6 +3,7 @@ package com.example.barrier_free.global.common;
 import java.util.List;
 
 import com.example.barrier_free.domain.place.enums.ImageType;
+import com.example.barrier_free.domain.place.enums.PlaceType;
 import com.example.barrier_free.domain.review.entity.Review;
 
 public interface Place {
